#assistants: move guideline headers into include folder
