+++ toonz/sources/include/tools/assistants/guidelineellipse.h
