+++ toonz/sources/include/tools/assistants/guidelineline.h
